Rename LineIterator to ThickLineIterator
