+++ embedded-graphics/src/primitives/mod.rs
@@ -3,6 +3,7 @@
 pub mod circle;
 pub mod line;
 pub mod rectangle;
+mod thick_line_iterator;
 pub mod triangle;
 
 pub use self::{circle::Circle, line::Line, rectangle::Rectangle, triangle::Triangle};
@@ -11,6 +12,7 @@ use crate::{
     pixelcolor::PixelColor,
     style::{PrimitiveStyle, Styled},
 };
+pub(crate) use thick_line_iterator::ThickLineIterator;
 
 /// Primitive trait
 pub trait Primitive: Dimensions {
